+++ packages/test-renderer/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @react-three/test-renderer
 
+## 7.0.22
+
+### Patch Changes
+
+- 30d38b1: remove logs
+- Updated dependencies [30d38b1]
+  - @react-three/fiber@7.0.23
+
+## 7.0.21
+
+### Patch Changes
+
+- 259e1fa: add camera:manual
+- Updated dependencies [259e1fa]
+  - @react-three/fiber@7.0.22
+
 ## 7.0.20
 
 ### Patch Changes